requirements.txt
streamlit==1.44.1
pandas==2.2.3
jinja2==3.1.2
xhtml2pdf==0.2.17
watchdog==3.0.0
zipfile36==0.1.3
python-slugify==8.0.1
Pillow==10.0.0  # Para manipulação de imagens